# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Jal Seva - Water Jar Delivery Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reports/daily:
    get:
      operationId: getDailyReport
      tags: [reports]
      summary: Get daily report
      parameters:
        - name: date
          in: query
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Daily report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DailyReport"
  /reports/monthly:
    get:
      operationId: getMonthlyReport
      tags: [reports]
      summary: Get monthly report
      parameters:
        - name: month
          in: query
          schema:
            type: integer
        - name: year
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Monthly report
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MonthlyReport"
...
components:
  schemas:
...
    CustomerMonthlySummary:
      type: object
      required: [customerId, customerName, month, year, totalJars, totalAmount, totalPaid, balance, deliveries, payments]
      properties:
        customerId:
          type: integer
        customerName:
          type: string
        month:
          type: integer
        year:
          type: integer
        totalJars:
          type: integer
        totalAmount:
          type: number
        totalPaid:
          type: number
        balance:
          type: number
        deliveries:
          type: array
          items:
            $ref: "#/components/schemas/Delivery"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
    Delivery:
      type: object
      required: [id, customerId, customerName, customerType, jarCount, amount, isPaid, deliveryDate, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        customerType:
          type: string
          enum: [people, shop]
        jarCount:
          type: integer
        amount:
          type: number
        isPaid:
          type: boolean
        deliveryDate:
          type: string
          format: date
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    Payment:
      type: object
      required: [id, customerId, customerName, amount, mode, paymentDate, createdAt]
      properties:
        id:
          type: integer
        customerId:
          type: integer
        customerName:
          type: string
        amount:
          type: number
        mode:
          type: string
          enum: [cash, upi, bank, other]
        paymentDate:
          type: string
          format: date
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
...
    DailyReport:
      type: object
      required: [date, totalDeliveries, totalJars, totalCash, totalUdhar, deliveries, payments]
      properties:
        date:
          type: string
        totalDeliveries:
          type: integer
        totalJars:
          type: integer
        totalCash:
          type: number
        totalUdhar:
          type: number
        deliveries:
          type: array
          items:
            $ref: "#/components/schemas/Delivery"
        payments:
          type: array
          items:
            $ref: "#/components/schemas/Payment"
    MonthlyReport:
      type: object
      required: [month, year, totalJars, totalAmount, totalCollected, totalOutstanding, totalCustomers, customerSummaries]
      properties:
        month:
          type: integer
        year:
          type: integer
        totalJars:
          type: integer
        totalAmount:
          type: number
        totalCollected:
          type: number
        totalOutstanding:
          type: number
        totalCustomers:
          type: integer
        customerSummaries:
          type: array
          items:
            $ref: "#/components/schemas/CustomerMonthlySummary"
...