# lib/api-zod/src/generated/types/dailyReport.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Jal Seva - Water Jar Delivery Management API
 * OpenAPI spec version: 0.1.0
 */
import type { Delivery } from './delivery';
import type { Payment } from './payment';

export interface DailyReport {
  date: string;
  totalDeliveries: number;
  totalJars: number;
  totalCash: number;
  totalUdhar: number;
  deliveries: Delivery[];
  payments: Payment[];
}